Updated game detection & interface initializing
Updated game detection & separated code used for getting the game code in 2 methods (GBA & DS).

Updated the interface initialization.
In some cases, the game code is only available after a few frames. In that case we retry a few times to show the proper interface instead of "Game not recognized!".

This is especially true for DS games where the DS bootup screen prevents retrieving the game code for 3 ~ 4 frames.