Corrected graphical issue, added method for creating buttons & interface adjustments
Fixed an issue where interface elements were missing padding / margin because the font and auto scale mode weren't set for the interface window itself.

Added a simplified method for generating buttons.

Changed where the FormBoderStyle is defined so that it's always set.

Made a property for the default font to keep things simpler.