Added Level & EXP for Bok 2
- Added memory addresses for Level & EXP
- Added Level & EXP to the main window
- Added these to the status edit subwindow too
- Added a class for representing U32 memory addresses values
- Updated method for setting status values to take U32 mem addresses values into account

Note : Level might need to be applied twice sometimes because the game takes too long to set the new EXP value, and thus adjust the Level based on the previous EXP for a few frames.
This is not an issue with the order of NumericUpDown fields or the memory addresses values instances themselves, but clearly with the game itself.