MVL Tool refactoring & assembly reference
- Added an assembly reference to C Sharp in the csproj file (required for dynamic type).
- Refactored the Memory Values Listing tool.

We now use the Bok Interface's instances for the current game's memory addresses instead of creating a new instance everytime the tool is instanciated.

Overall instanciating the tool is now much easier, and some parameters that should never change are directly set in the class itself (for example its subwindow title).