Update regarding maximum item durability
- Updated the "Tomato Juice" Item instance for Bok 2, as that item is perishable.
- Added code in the ZoktaiInventoryEditor class for adjusting the maximum durability that can be set for a slot based on the selected item.

Depending on the selected item, the maximum durability that can be set can be different.
The maximum is now updated when selected another item for the slot, or when the item is being retrieved from the current in-game inventory.

By default, the highest maximum value is set (this prevents a crash when the current in-game inventory is retrieved and an item's durability value is high).