Added Accessory & ZoktaiAccessories classes
- Added ZoktaiAccessories class, which is used to generate & hold a list of Accessory instances.
- Added Accessory class, which represents an accessory.

In Boktai games, armors are refered to as "Accessories" (or "Protectors" in Bok 2's case).

In Bok 2, these are always for the Torso, as the game has no other slot for equipping armors.

In Bok 3 & DS, each accessory has a type used to define which slot it can go to (Head, Torso, Hand or Foot).

In Bok 2, armors also have Durability (defensive value) & Weight (affects movement speed based on Django's AGI stat). This isn't the case in the other games.