Refactoring regarding Items-related classes
- Turned the Item class into an abstract class. Each game should now have a class for its items that inherits from the Item class.

- Added the ZoktaiItem class for representing an Item in Bok 2.

- Updated the ZoktaiItems class to instanciate ZoktaiItem instances.

This change is necessary because of some things being slightly different between games.

For example, the value for an item is different between Bok 2 & 3. This means that the checks we used to do for setting the rottenAt & rottsInto properties cannot be the same.

The changes in this commit, although improvable, enables us to instanciate Items while also being able to adjust for these slight differences easily.