Update MemoryValues.cs
- Changed how classes containing memory addresses are initialized when filling MemoryValues dictionnaries.

The memory addresses classes were previously in properties of the MemoryValues class, which meant that they were all initialized even when not used (only the one for the current game will actually be used).

The properties have been removed and instead the class containing memory addresses for the current game will be initialized only when filling MemoryValues dictionnaries.

This can still be improved by might require refactoring other files.