Fixed crash when closing ROM
- Fixed an issue where closing a ROM could cause the Bok Interface to crash due to memory domains being unavailable.

For example during the initialization of the Bok 2 interface, we read the memory to determine which version is being loaded.

Additionaly, some BokInterface properties not being reset when closing a ROM would lead to the above error. By resetting the proper properties the code causing this error will not be reached.